Move TripIdAndServiceDate to transit model
